requirements.txt
Flask
SQLAlchemy
feedparser
beautifulsoup4
Flask-Script
tabulate
requests_oauthlib
emoji